+++ expenses/management/commands/populate_local_amount.py
@@ -1,0 +1,14 @@
+from django.core.management.base import BaseCommand
+from expenses.models import Expense
+
+
+class Command(BaseCommand):
+    help = "Populate the expenses' local amounts"
+
+    def handle(self, *args, **options):
+        for expense in Expense.objects.exclude(local_amount=0, currency__alpha3="HND"):
+            expense.local_amount = expense.get_local_amount
+            expense.save()
+            print(f"Updated expense {expense.id}")
+        print("Done!")
+        return 0